deletion: delete chunk overlapping multiple tables when processing oldest table that has it indexed (#11600)
**What this PR does / why we need it**:
When a delete request has a line filter and/or is deleting a chunk
partially, we create a new chunk from the source chunk and mark the
original chunk for deletion. However, when a chunk overlaps multiple
tables, we apply heuristics to retain the chunk until all the tables
that have it indexed get processed since we need the original chunk to
build a new one.

Earlier, we used to process tables from oldest to newest, so the logic
to detect if all the tables indexing a chunk are processed was done by
seeing if `chunk-end-time` <= `table-end-time`. To put it in simple
terms, if the chunk ends in the current table we are processing, it
means this is the last table that we are processing, which has the chunk
indexed.

However, this does not work anymore since we changed the order of
processing tables to process the newest tables first. To get it working
back, I have updated the logic to `chunk-start-time` >=
`table-start-time`. In simple terms, if the chunk starts in the current
table we are processing, it means this is the last table we are
processing, which has the chunk indexed.

**Checklist**
- [x] Tests updated