fix: remove shared state noParserHints to avoid data race (#11685)
**What this PR does / why we need it**:

A data race introduced in https://github.com/grafana/loki/pull/11587 was
caught in the [backport to
k184](https://github.com/grafana/loki/pull/11668). This removes the
shared state of a single global `noParserHints` in favor of creating an
empty `Hint` object for each label builder, since the `Hints` is keeping
state of `extracted` and `requiredLabels`.