Remove unnecessary wrapper around bloom shipper
The bloom store was initially thought as a translation layer between the
bloom gateway and the bloom shipper to be able to simplify the API of
the shipper.

However, it turned out that the store calls were only a 1-to-1 mapping
of the shipper calls and therefore adding unnecessary complexity and
call stack.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>