Bloom Gateway: Partition requests into multiple tasks
This PR changes at which point a request is split into multiple tasks in
case the chunks boundaries of the request overlap multiple days.

This solves the problem that the request handler needs to know for how
many responses from the block querier needs to wait.

Previously, a single task that spanned across multiple days was executed
multiple times (once for each matching day).

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>