feat: add release please config (#11333) (#11789)
**What this PR does / why we need it**:

The new release process is being developed in another repo for now as it
enables much faster feedback loops. We are at the stage where we are
pulling the Loki repo and building images, so it's time to move the
`release-please` config into Loki, and configure it for the various
packages we will be releasing.

This PR adds the two basic configs that `release-please` requires. This
PR does not change anything about the release process, it just enables
us to keep iterating out of band until we're ready to move over.

This was already added to `main`, but adding here so I can test it on
the release branch

**Which issue(s) this PR fixes**:
Re #11332