[Bloom] Keyspace aware bloom creation, library utilities (#11810)
This PR continues work refactoring the bloom compactor pkg and adds a
bunch of utilities for keyspace manipulation/mgmt. A lot of the bloom
block logic is contained here as a skeleton.

This code is not complete, but is fine to merge as we don't use this yet
and should be easier to review before it gets too large.