Implement block fetching in BloomStore fetcher
The fetcher implements a new function `FetchBlocks()` which returns a
list of block directories.

A block directory represents a local file path that contains the
extracted files of a bloom block. It also holds a counter of active
readers that access the directory. This is used for safely deleting the
directory in case it needs to be removed from disk (eg. max disk/cache
size reached).

The fetcher resolves the block directory from three places:
1. Cache: The in-memory cache that holds actively accessed directories
   and keeps track of used disk size.
2. Filesystem: In case the cache was emptied (e.g. when restarting the
   process) but the block directory is present on disk, it can be
   re-assambled into a bloom directory that can be put to cache.
3. Storage: If the directory is not present locally, the block archive
   is downloaded and extracted and the block directory is put to cache.