perf(blooms): Recycle bloom page buffers (#12256)
Every time when reading a bloom page into memory, a new byte slice with the size of the decoded page was allocated.

This PR changes bloom page decoder to use a byte slice from the `BlockPool`. The caller (`LazyBloomIter` of the block querier) is responsible for closing the page whenever a new one is loaded, so the byte slice is put back to its pool. 

Additionally, this PR changes the max bucket of the `BlockPool` byte slice pool to 64MB so that pages bigger than 16MB are correctly put back to the pool, since [slices exceeding the capacity of the largest bucket are not put back](https://github.com/grafana/loki/blob/main/vendor/github.com/prometheus/prometheus/util/pool/pool.go#L74-L87).

---
Signed-off-by: Christian Haudum <christian.haudum@gmail.com>