Remove bloom gateway ring
Remove ring in favour of a client side, DNS discovery of the bloom
gateway servers combined with consistent hashing using the jumphash
algorithm.

Instead of configuring the ring, the bloom gateway client needs to be
configured with the `-bloom-gateway-client.addresses` flag, which takes
a list of comma separated, DNS-discovery-style strings.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>