fix(blooms): Correctly return unfiltered chunks for series that are not mapped to any block (#12774)
This PR fixes a conceptual mistake in the code of resolving blocks on the index gateways. Currently, a series does not resolve to any block is discarded instead of being kept for the response.
This change adds the chunks of the skipped series to the bloom querier response.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>
(cherry picked from commit c36b1142c7acd6a13a3634ddbef71254040cff73)