fix(blooms): Fully deduplicate chunks from FilterChunkRef responses (#12807)
This PR aims for full de-duplication of chunks and series from filter requests from the index gateway to the bloom gateway.

Whenever we merge/de-duplicate slices, the inputs need to be sorted. It appears that the Removals (chunks) from the v1.Output are not guaranteed to be sorted.

When comparing ShortRefs, both From, Through, and Checksum need to be used.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>
(cherry picked from commit a0f358fcc8295d93ee38b67738e8d90045c50dab)