Handle not found metas gracefully
There is a time window between between listing metas and fetching them
from object storage which could lead to a race condition that the meta
is not found in object storage, because it was deleted and superseded by
a newer meta.

This can happen when querying recent bloom data, that is still subject
to updates.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>