fix(blooms): bloomshipper no longer returns empty metas on fetch (#13130)
Traced some bloom compactor problems to this store interface returning empty `Metas` (with the exception of the reference key) on fetch when the object does not exist in storage. I originally found this due to a panic in the outdated meta removal code because empty `Metas` don't have any source TSDBs.

Fixes the impl to only return metas which existed which have their contents populated.