refactor: Remove unnecessary spanlogger usage (#13255)
Historically, we've been using `spanlogger` to enrich our spans. That isn't the right usage for it (as of now), as they'll be emitting log lines. Our current usage is causing:
- Noisy/important log lines less visible
- Unnecessary logging volume
- Spread of the bad practice.
With this PR I'm moving away from `spanlogger` all places where the log message is too bad or nonexistent. That's because these cases are indicating we don't care about that log line at all, and only about the data injected in the span.