feat: Fail request when log selector expression is used as instant query
A log selector expression is a LogQL expression for returning logs, in
contrast to sample expressions, which return metrics. The simplest form
of log selector expressions are label matchers, e.g. `{env="prod"}`.

This PR changes the behaviour of Loki so it does not allow using a log
selector expression when the instant query endpoint `/api/v1/query` is
queried.

Previously this endpoint allowed these types of log queries, but returns
inconsistent results, which where a major cause for confusion. Returning
a concise error helps the user understand that they likely selected the
wrong query type in Grafana when executing the query.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>