fix: Fix panic in BloomStore initialisation
The initialisation of the `BloomStore` service failed with a panic if
the index gateways were configured to use `ring` mode.

This happened, because the bloom store requires to wire up the
configuration for the shipper store, see
https://github.com/grafana/loki/blob/1a508b0aa/pkg/loki/modules.go#L864-L867

At this time, the index gateway ring manager wasn't initialized yet,
leading to a `nil` pointer dereference.

This PR fixes the issue by adding the index gateway ring as dependency
of the bloom store so it gets initialised first.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>