fix(blooms): Improve error wrap to make ignoreNotFound work when fetching blocks (#13656)
Fix an issue where the `options.ignoreNotFound` of the fetcher is not effective. 

The root cause of this issue is that the error returned by `fmt.Errorf()` in `BloomClient` loses the underlying cause. As a result, `S3ObjectClient.IsObjectNotFoundErr()` fails to determine if the underlying cause of the returned error is of type `awserr.Error`.