Use mark-sweep to delete WAL files older than retention period
This commit updates the metastore to use mark and sweep to delete
segments older than the retention period. A mark and sweep is
initated by the leader in the Raft quorum via the periodic
maintenance. It marks all segments older than the retention period
and replicates this command to all followers, and sweeps marked
segments older than the grace period, deleting them from object
storage and replicating the command to all followers too. While
updates to each instance's local boltdb is replicated to ensure
it is consistent on all replicas, the object storage operations are
not replicated and are only performed by the leader in the quorum.