chore(blooms)!: Remove bloom compactor component (#13969)
**What this PR does / why we need it**:

This commit removes the code related to the bloom compactor which is superseded by the bloom planner and builders.

A handful of CLI arguments of per-tenant settings changed their prefix from `-bloom-compactor.*` to `-bloom-build.*`. Other settings for the compactor component itself, also prefixed with `-bloom-compactor.*` were removed. See upgrade docs for further information.

**Special notes for your reviewer**:

:heavy_check_mark: Part of https://github.com/grafana/loki/issues/13957

:notebook_with_decorative_cover: Documentation update https://github.com/grafana/loki/pull/13965

:warning: Integrating the bloom planner and builder into the `backend` target is done in a follow-up PR.

---

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>