feat: add retries for s3 ObjectExists calls (#14062)
**What this PR does / why we need it**:
It was determined via careful examination that Loki would issue retries for S3 `GetObject` calls, but would not retry calls to `ObjectExists`.

For transient issues (e.g. rate-limiting), it makes sense to retry the command to see if an object does exist.  If the object is not found (meaning, a 404 return code, and a successful query of the S3 storage to show the object does not exist), the retries are not leveraged, as it is not needed.