storage/bloom/v1: filter out chunks where the queried key was not indexed
Previously, this logic assumed that once we're running the bloom test,
any key that was not indexed may still be in the chunk but not in the
bloom. However, as long as the bloom exists, we can be confident that if
the key does not exist, the key will not be found anywhere in that
chunk.