Feat: Add kubeVersionOverride for Helm chart
With this change is now possible to override the
`.Capabilities.KubeVersion.Version` parameter with a fixed Kubernetes
version using `kubeVersionOverride` in values.yaml file.

Overriding the kubeVersion allows chart's user to generate templates
from the chart that can be deployed to cluster that are not directly
accessible from the user's CLI and run a different Kubernetes version.

Signed-off-by: Manuel Castellin <manuel@castellinconsulting.com>