chore(bloom): support simplifiable regexp matchers
This adds support for basic regexps which can be simplified into a
sequence of OR matchers, such as:

* `key=~"value"` becomes `key="value"`
* `key=~"value1|value2"` becomes `key="value1" or key="value2"`.

Matchers like `key=~".+"` continue to not be supported because the lack
of a key doesn't mean that it doesn't exist as a label.

Only the cases above are "officially" supported. However, we technically
support basic concatenations and character classes due to how
regexp/syntax parses and simplifies expressions such as `value1|value2`
into `value[12]`.

To prevent unbounded cardinality, we limit regexp expansion to 25
matchers; otherwise a regexp like `value[0-9][0-9][0-9][0-9]` would
expand into 10,000 matchers (too many!).