feat(lambda-promtail): allow prefix matching for CloudWatch log groups
If you have a lot of log groups with similar names this is much easier than listing them all explicitly. Or the bonus feature of just including them all by specifying a prefix of "".

While we're here, introduce the ability to add filter patterns on the subscription to narrow the matched logs as desired.