Fix regression in bloom gateway that caused nothing to be filtered
Pull request #14661 added the series labels to the FilterChunkRefs
request in order to be able to filter out full series that do not
contain the filter key when using bloom filters.

However, this chage incorrectly assumed that `Chunk.Metric` from the
`GetChunks()` call contains the labels of the stream. This information
is only populated once the chunk is fetched, though. Therefore the list
of labels was empty, and no chunk refs were filtered.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>