Remove `wget` from Promtail docker image
The package has been added to the Docker image with PR #11711 with the
intention to support the Docker healthcheck.

However, to reduce the attack surface of our Docker images, we want to
keep them as slim as possible. The current version of Promtail (3.3.0)
for example contains a wget version with vulnerability
[CVE-2024-38428](https://security-tracker.debian.org/tracker/CVE-2024-38428).

The healthcheck can be achieved by other means, e.g.

1. Extend the `grafana/promtail` base image and add `wget` using `apt
   install wget`
   https://github.com/grafana/loki/issues/11590#issuecomment-2106037181
2. Use low-level `/dev/tcp/127.0.0.1:9080` to establish a connection and
   check the exit code
   https://github.com/grafana/loki/issues/11590#issuecomment-2333481326

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>