feat: remove target lag and keep just maximum lag
This commit is an experiment to see if we can remove target lag and
just keep maximum lag. The goal here is to simplify the code and
make the catch up feature easier to reason about. However, if we find
that this causes stale reads, we can add it back.