fix(blooms): Fix check for skipping most recent data when filtering blooms (#15300)
After changing the default of  `-bloom-build.planner.min-table-offset` from `1` (yesterday) to `0` (today), we noticed that the bloom gateways reported a very high percentage of missing series in blocks.

This is because the gateway received filter requests for today's block, but the requested series is too new and has not been added to the bloom block yet.

In the bloom querier on the index gateway there is a condition under which requests to the bloom gateway are skipped, because the blooms are likely not available yet.

This PR changes this condition to correctly take the min-table-offset as well as the planning-interval into account.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>
(cherry picked from commit 78d3c4415662b4003fc275b29aff053d531d90f0)