Update docs on using terraform to create s3 buckets for accuracy
The oidc_provider variable is invalid for the Terraform module, which extracts the provide info from the given EKS cluster.
The terraform requires a namespace for the service account, which was undocumented here.
The s3 bucket name will not work as loki-data as s3 bucket names are global, so we should encourage setting it explicitly.
The syntax for applying a terraform module was incorrect: missing `apply` command and variables were given with incorrect syntax.