feat(api): Support Parquet as a query response format. (#15408)
**What this PR does / why we need it**:
This changes adds support for the header `Accept: application/vnd.apacher.parquet`. If this header is set the response for metric and stream queries will response with a Parquet file.

A metrics response has the columns `timestamp`, `labels` and `value`. A stream response has the columns `timestamp`, `labels` and `line`.

Co-authored-by: Christian Haudum <christian.haudum@gmail.com>